moving helper functions to functions.rs
